Create Absence fixed.

+++ Framework.AuthServer/Dtos/AbsenceService/Output/AbsenceDTO.cs
@@ -10,7 +10,7 @@ namespace Framework.AuthServer.Dtos.AbsenceService.Output
         public required string Description { get; set; }
         public required AbsenceTypes Type { get; set; }
         public required AbsenceStatus Status { get; set; }
-        public required short Duration { get; set; }
+        public required double Duration { get; set; }
         public DateTime StartTime { get; set; }
         public DateTime EndTime { get; set; }
     }
